+++ NEWS.md
@@ -1,2 +1,10 @@
 # RPESE 1.0.0
 * Initial stable release of package.
+
+# RPESE 1.0.1
+* Fix bug for SemiSD.
+* Add functionality of package for frequency decimation and truncation.
+
+# RPESE 1.0.2
+* Add the option to return the coefficients from the Exponential/Gamma fit.
+